services/etf_service.py
from typing import Dict, List, Optional
import logging
from datetime import datetime
import requests
from bs4 import BeautifulSoup
from tenacity import retry, stop_after_attempt, wait_exponential
import aiohttp

logger = logging.getLogger(__name__)

class ETFService:
    def __init__(self):
        self.base_url = "https://www.twse.com.tw"
        self.headers = {
            'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36'
        }

    async def analyze_etf(self, etf_code: str) -> str:
        """分析ETF"""
        try:
            holdings = await self.get_etf_holdings(etf_code)
            if not holdings:
                return f"無法獲取 {etf_code} 的成分股資訊"

            # 基本分析
            analysis = f"===== {etf_code} ETF分析 =====\n"
            analysis += f"成分股數量: {len(holdings)}\n"
            # 其他分析邏輯...
            
            return analysis
        except Exception as e:
            logger.error(f"分析ETF時發生錯誤: {str(e)}")
            return f"分析 {etf_code} 時發生錯誤"

    @retry(stop=stop_after_attempt(3), wait=wait_exponential(multiplier=1, min=4, max=10))
    async def get_etf_holdings(self, etf_code: str) -> List[str]:
        """獲取ETF持股"""
        try:
            url = f"{self.base_url}/zh/page/ETF/list.html"
            async with aiohttp.ClientSession() as session:
                async with session.get(url, headers=self.headers) as response:
                    text = await response.text()
                    
            # 模擬返回測試數據
            if etf_code == "0050":
                return ["2330", "2317", "2454", "2412", "2308"]
            return []
            
        except Exception as e:
            logger.error(f"獲取ETF持股時發生錯誤: {str(e)}")
            return []

etf_service = ETFService()
