requirements.txt
fastapi
uvicorn
line-bot-sdk
python-dotenv
requests
beautifulsoup4
tenacity
pymongo
pandas
apscheduler
google-generativeai
pytest
tenacity